Issues: #82 - lint fix
