+++ lib/frontend/src/actions/index.js
@@ -2,10 +2,10 @@ import equal from 'deep-equal'
 import fetch from 'isomorphic-fetch'
 import keyIndex from 'react-key-index'
 
+import { fetchIssues } from './issues'
 import {
   allLabelsFromState,
   getMissingAndOverlyLabels,
-  querystringize,
   reportRangeFromState,
   repositoryNameFromState,
   timelineRangeFromState,
@@ -78,7 +78,7 @@ export const updateRepoLabels = missingAndOverlyLabels => ({
   missingAndOverlyLabels,
 })
 
-const stateToParams = (state, target) => {
+export const stateToParams = (state, target) => {
   let users
   switch (target) {
     case 'tickets':
@@ -225,79 +225,6 @@ export const fetchRepository = () => async (dispatch, getState) => {
     dispatch(changeRepository(repositoryJson.objects, labelsJson.objects))
     return dispatch(setMissingAndOverlyLabels())
   }
-}
-
-export const fetchIssues = () => async (dispatch, getState) => {
-  let response, json
-  let ticketsJson = {},
-    commentsJson = {}
-  const state = getState()
-  const params = querystringize(stateToParams(state, 'tickets'))
-  try {
-    response = await fetch(`/api/tickets?${params}`, {
-      method: 'GET',
-      credentials: 'same-origin',
-    })
-  } catch (e) {
-    return dispatch(
-      setError(
-        'Erreur pendant la récupération des données des tickets',
-        'repository'
-      )
-    )
-  }
-  if (response.status >= 200 && response.status < 300) {
-    try {
-      ticketsJson = await response.json()
-    } catch (e) {
-      return dispatch(setError('La réponse ne contient pas de json', 'issues'))
-    }
-  } else {
-    const jsonMessage = typeof json === 'undefined' ? '' : json.message
-    return dispatch(
-      setError(
-        `${response.status}: ${response.statusText} ${jsonMessage}`,
-        'repository'
-      )
-    )
-  }
-  for (const index in ticketsJson.objects) {
-    if (ticketsJson.objects.hasOwnProperty(index)) {
-      if (ticketsJson.objects[index].nb_comments > 0) {
-        try {
-          response = await fetch(
-            `/api/repos/${ticketsJson.objects[index].repo_name}/tickets/${
-              ticketsJson.objects[index].ticket_number
-            }/comments`,
-            {
-              method: 'GET',
-              credentials: 'same-origin',
-            }
-          )
-        } catch (e) {
-          return dispatch(
-            setError(
-              'Erreur pendant la récupération des données des commentaires',
-              'issues'
-            )
-          )
-        }
-        if (response.status >= 200 && response.status < 300) {
-          try {
-            commentsJson = await response.json()
-          } catch (e) {
-            return dispatch(
-              setError('La réponse ne contient pas de json', 'issues')
-            )
-          }
-          ticketsJson.objects[index].comments = commentsJson.objects
-        }
-      } else {
-        ticketsJson.objects[index].comments = []
-      }
-    }
-  }
-  return dispatch(setResults({ ['issues']: ticketsJson.objects }, 'issues'))
 }
 
 export const updateATicket = (ticket, values) => async dispatch => {
