+++ lib/frontend/src/actions/issues.js
@@ -1,0 +1,91 @@
+import fetch from 'isomorphic-fetch'
+
+import { stateToParams, setError, setResults } from './index'
+import { querystringize } from '../utils'
+
+// specific error until all merge
+export const setErrorIssue = errors => ({
+  type: 'SET_ERROR_ISSUE',
+  errors,
+})
+
+export const fetchIssues = () => async (dispatch, getState) => {
+  let response, json
+  let ticketsJson = {},
+    commentsJson = {}
+  let allErrors = ''
+  const state = getState()
+  const params = querystringize(stateToParams(state, 'tickets'))
+  try {
+    response = await fetch(`/api/tickets?${params}`, {
+      method: 'GET',
+      credentials: 'same-origin',
+    })
+  } catch (e) {
+    return dispatch(
+      setError(
+        'Erreur pendant la récupération des données des tickets',
+        'repository'
+      )
+    )
+  }
+  if (response.status >= 200 && response.status < 300) {
+    try {
+      ticketsJson = await response.json()
+    } catch (e) {
+      return dispatch(setError('La réponse ne contient pas de json', 'issues'))
+    }
+  } else {
+    const jsonMessage = typeof json === 'undefined' ? '' : json.message
+    return dispatch(
+      setError(
+        `${response.status}: ${response.statusText} ${jsonMessage}`,
+        'repository'
+      )
+    )
+  }
+  for (const index in ticketsJson.objects) {
+    if (ticketsJson.objects.hasOwnProperty(index)) {
+      if (ticketsJson.objects[index].nb_comments > 0) {
+        try {
+          response = await fetch(
+            `/api/reposs/${ticketsJson.objects[index].repo_name}/tickets/${
+              ticketsJson.objects[index].ticket_number
+            }/comments`,
+            {
+              method: 'GET',
+              credentials: 'same-origin',
+            }
+          )
+        } catch (e) {
+          allErrors.push(
+            `Erreur pendant la récupération des commentaires du ticket ${
+              ticketsJson.objects[index].ticket_number
+            }`
+          )
+        }
+        if (response.status < 200 || response.status >= 300) {
+          allErrors +=
+            'Erreur pendant la récupération des commentaires du' +
+            ` ticket ${ticketsJson.objects[index].ticket_number}, `
+          ticketsJson.objects[index].comments = []
+          continue
+        }
+        try {
+          commentsJson = await response.json()
+        } catch (e) {
+          allErrors +=
+            'Erreur pendant la récupération des commentaires du' +
+            ` ticket ${ticketsJson.objects[index].ticket_number}, `
+        }
+        ticketsJson.objects[index].comments = commentsJson.objects
+      } else {
+        ticketsJson.objects[index].comments = []
+      }
+    }
+  }
+  dispatch(setResults({ ['issues']: ticketsJson.objects }, 'issues'))
+  if (allErrors !== '') {
+    dispatch(setErrorIssue(allErrors.replace(/, $/g, '')))
+  }
+}
