+++ lib/frontend/src/reducer/index.js
@@ -102,6 +102,13 @@ const issues = (state = initial.issues, action) => {
           }),
         },
       }
+    case 'SET_ERROR_ISSUE':
+      return {
+        results: {
+          ...state.results,
+        },
+        error: action.errors,
+      }
     default:
       return handleResultsLoadingAndError(state, 'issues', action)
   }
